Fixed Favorites

+++ Backend/app/Http/Controllers/UserControllers/FavoriteController.php
@@ -35,45 +35,15 @@ class FavoriteController extends Controller
         $validated = $request->validate([
             'medicine_id' => 'required'
         ]);
-        $user =  auth('sanctum')->user();
+        $user = auth('sanctum')->user();
         if ($user == null) {
             return response()->json([
                 'message' => 'Invalid User'
             ], 400);
         }
-        // Get user id
-        $user_id = $user->id;
-        $favorite = Favorite::where('user_id', $user_id)->first();
+        // Get user Favorite
+        $favorite = Favorite::where('user_id', $user->id)->first();
         // Check medicines
-            $medicine_id = $validated['medicine_id'];
-            $medicine = Medicine::find($medicine_id);
-            if ($medicine == null) {
-                return response()->json([
-                    'message' => 'Invalid Medicine'
-                ], 400);
-            }
-        $favorite->Medicines()->attach($medicine);
-// Response
-return response()->json([
-    'message' => 'Successfully Added To Favorite'
-], 200);
-    }
-
-    //Removing The Medicine From Favorites
-public function DeleteFavorite(Request $request){
-    $validated = $request->validate([
-        'medicine_id' => 'required'
-    ]);
-    $user =  auth('sanctum')->user();
-    if ($user == null) {
-        return response()->json([
-            'message' => 'Invalid User'
-        ], 400);
-    }
-    // Get user id
-    $user_id = $user->id;
-    $favorite = Favorite::where('user_id', $user_id)->first();
-    // Check medicines
         $medicine_id = $validated['medicine_id'];
         $medicine = Medicine::find($medicine_id);
         if ($medicine == null) {
@@ -81,13 +51,48 @@ public function DeleteFavorite(Request $request){
                 'message' => 'Invalid Medicine'
             ], 400);
         }
-    $favorite->Medicines()->detach($medicine);
-// Response
-return response()->json([
-'message' => 'Successfully Removed From Favorite'
-], 200);
+        $medicines = $favorite->Medicines;
+        foreach($medicines as $medicine1) {
+            if($medicine['id'] == $medicine1['id']){
+                return response()->json([
+                    'message' => 'Medicine Already In Favorites'
+                ], 200);
+            }
+        }
+        $favorite->Medicines()->attach($medicine);
+        return response()->json([
+            'message' => 'Successfully Added To Favorite'
+        ], 200);
+    }
 
+    //Removing The Medicine From Favorites
+    public function DeleteFavorite(Request $request)
+    {
+        // Validate request
+        $validated = $request->validate([
+            'medicine_id' => 'required'
+        ]);
+        $user = auth('sanctum')->user();
+        if ($user == null) {
+            return response()->json([
+                'message' => 'Invalid User'
+            ], 400);
+        }
+        // Get user Favorite
+        $favorite = Favorite::where('user_id', $user->id)->first();
+        // Check medicines
+        $medicine_id = $validated['medicine_id'];
+        $medicine = Medicine::find($medicine_id);
+        if ($medicine == null) {
+            return response()->json([
+                'message' => 'Invalid Medicine'
+            ], 400);
+        }
+        $favorite->Medicines()->detach($medicine);
+        // Response
+        return response()->json([
+            'message' => 'Successfully Removed From Favorite'
+        ], 200);
     }
 
 }
-
